WIP: Solve build errors
