+++ app/src/main/java/com/ivy/wallet/ui/home/HomeEvent.kt
@@ -1,8 +1,8 @@
 package com.ivy.wallet.ui.home
 
+import com.ivy.core.data.model.TimePeriod
 import com.ivy.wallet.domain.data.core.Transaction
 import com.ivy.wallet.domain.deprecated.logic.CustomerJourneyCardData
-import com.ivy.wallet.ui.onboarding.model.TimePeriod
 
 sealed class HomeEvent {
     object Start : HomeEvent()
