+++ app/src/main/java/com/ivy/wallet/ui/home/HomeMoreMenu.kt
@@ -51,17 +51,17 @@ import com.ivy.design.l0_system.Theme
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
 import com.ivy.frp.view.navigation.navigation
-import com.ivy.wallet.Constants
+import com.ivy.navigation.BudgetScreen
+import com.ivy.navigation.Categories
+import com.ivy.navigation.Loans
+import com.ivy.navigation.PlannedPayments
+import com.ivy.navigation.Report
+import com.ivy.navigation.Search
+import com.ivy.navigation.Settings
 import com.ivy.resources.R
-import com.ivy.wallet.ui.BudgetScreen
-import com.ivy.wallet.ui.Categories
+import com.ivy.wallet.Constants
 import com.ivy.wallet.ui.IvyWalletPreview
-import com.ivy.wallet.ui.Loans
-import com.ivy.wallet.ui.PlannedPayments
-import com.ivy.wallet.ui.Report
 import com.ivy.wallet.ui.RootActivity
-import com.ivy.wallet.ui.Search
-import com.ivy.wallet.ui.Settings
 import com.ivy.wallet.ui.donate.DonateScreen
 import com.ivy.wallet.ui.ivyWalletCtx
 import com.ivy.wallet.ui.theme.Blue
