+++ app/src/main/java/com/ivy/wallet/ui/home/HomeState.kt
@@ -1,14 +1,14 @@
 package com.ivy.wallet.ui.home
 
+import com.ivy.core.IvyWalletCtx
+import com.ivy.core.data.model.TimePeriod
 import com.ivy.design.l0_system.Theme
 import com.ivy.wallet.domain.data.TransactionHistoryItem
 import com.ivy.wallet.domain.deprecated.logic.CustomerJourneyCardData
 import com.ivy.wallet.domain.pure.data.IncomeExpensePair
-import com.ivy.wallet.ui.IvyWalletCtx
 import com.ivy.wallet.ui.data.AppBaseData
 import com.ivy.wallet.ui.data.BufferInfo
 import com.ivy.wallet.ui.data.DueSection
-import com.ivy.wallet.ui.onboarding.model.TimePeriod
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import java.math.BigDecimal
