+++ app/src/main/java/com/ivy/wallet/ui/home/HomeViewModel.kt
@@ -1,11 +1,17 @@
 package com.ivy.wallet.ui.home
 
+import com.ivy.core.IvyWalletCtx
+import com.ivy.core.data.model.MainTab
+import com.ivy.core.data.model.TimePeriod
+import com.ivy.core.data.model.toCloseTimeRange
 import com.ivy.design.l0_system.Theme
 import com.ivy.frp.fixUnit
 import com.ivy.frp.then
 import com.ivy.frp.thenInvokeAfter
 import com.ivy.frp.view.navigation.Navigation
 import com.ivy.frp.viewmodel.FRPViewModel
+import com.ivy.navigation.BalanceScreen
+import com.ivy.navigation.Main
 import com.ivy.wallet.domain.action.account.AccountsAct
 import com.ivy.wallet.domain.action.category.CategoriesAct
 import com.ivy.wallet.domain.action.exchange.SyncExchangeRatesAct
@@ -14,26 +20,25 @@ import com.ivy.wallet.domain.action.settings.CalcBufferDiffAct
 import com.ivy.wallet.domain.action.settings.SettingsAct
 import com.ivy.wallet.domain.action.settings.UpdateSettingsAct
 import com.ivy.wallet.domain.action.transaction.HistoryWithDateDivsAct
-import com.ivy.wallet.domain.action.viewmodel.home.*
+import com.ivy.wallet.domain.action.viewmodel.home.HasTrnsAct
+import com.ivy.wallet.domain.action.viewmodel.home.OverdueAct
+import com.ivy.wallet.domain.action.viewmodel.home.ShouldHideBalanceAct
+import com.ivy.wallet.domain.action.viewmodel.home.UpcomingAct
+import com.ivy.wallet.domain.action.viewmodel.home.UpdateAccCacheAct
+import com.ivy.wallet.domain.action.viewmodel.home.UpdateCategoriesCacheAct
 import com.ivy.wallet.domain.action.wallet.CalcIncomeExpenseAct
 import com.ivy.wallet.domain.action.wallet.CalcWalletBalanceAct
 import com.ivy.wallet.domain.data.core.Account
 import com.ivy.wallet.domain.data.core.Settings
 import com.ivy.wallet.domain.data.core.Transaction
+import com.ivy.wallet.domain.deprecated.logic.CustomerJourneyCardData
 import com.ivy.wallet.domain.deprecated.logic.CustomerJourneyLogic
 import com.ivy.wallet.domain.deprecated.logic.PlannedPaymentsLogic
 import com.ivy.wallet.domain.deprecated.logic.currency.ExchangeRatesLogic
-import com.ivy.wallet.domain.deprecated.logic.CustomerJourneyCardData
 import com.ivy.wallet.domain.pure.data.ClosedTimeRange
-import com.ivy.wallet.ui.BalanceScreen
-import com.ivy.wallet.ui.IvyWalletCtx
-import com.ivy.wallet.ui.Main
 import com.ivy.wallet.ui.data.AppBaseData
 import com.ivy.wallet.ui.data.BufferInfo
 import com.ivy.wallet.ui.data.DueSection
-import com.ivy.wallet.ui.main.MainTab
-import com.ivy.wallet.ui.onboarding.model.TimePeriod
-import com.ivy.wallet.ui.onboarding.model.toCloseTimeRange
 import com.ivy.wallet.utils.dateNowUTC
 import com.ivy.wallet.utils.ioThread
 import dagger.hilt.android.lifecycle.HiltViewModel
