+++ app/src/main/res/layout/widget_add_transaction.xml
