+++ settings.gradle.kts
@@ -16,4 +16,3 @@ include(":ivy-design")
 include(":ivy-core")
 include(":ivy-navigation")
 include(":ivy-resources")
-include(":widgets")
